fix evaluator
